Add ability to trace cable paths through circuits. (#104)
* Add ability to trace cable paths through circuits.

* Add test for single path through circuit.

* A cable should not be connected to both sides of a circuit.

Updating test to match reality and fix the loop.

* Update signals for cables to always call rebuild_paths, even on cable creation.

This handles the case where creating a path through a circuit doesn't update
the first cable, when a second cable is connected.  This mimics the same
action that is done when creating path through front/rear ports.  This allows
a circuit to be both a Path Endpoint for cases such as DSL or Fiber Broadband,
but also updates the cables when using the Circuit as a passthrough such as
dark fiber.

* Remove bug where adding the cable to the path twice.

* Correct tests for circuit paths.

* Add additional argument to create_cablepath to prevent automic database issues.

* Update docs to reflect refreshing the cable path.

* Make changes requested by reviewers.

* Circuit Terminations can now be added later without the User required to update paths manually.

* Add Circuit information to Interfaces "Connection" Column.

* Remove test variable.

Was demoing vscode devcontainers and I forgot to remove.

* Remove extra line for Jathan.

* Add device status to tests and remove fontAwesome.