Travis CICD Build Improvements (#211)
The purpose of this PR is to attempt to speed up the current build times. The original goal was not to break everything out but I'm testing speed differences. This PR also cleans up having tests defined in `cibuild.sh`, `tasks.py`, and pre-commit. I also added the ability to specify `INVOKE_LOCAL` to execute invoke tasks locally instead of in the docker containers if set to `True`, this allows the `cibuild.sh` script to execute the tests without building a docker container.

Co-authored-by: Glenn Matthews <glenn.matthews@networktocode.com>
Co-authored-by: Jathan McCollum <jathan@gmail.com>