Rack power utilization (#2646)
* Remove extra closing of tag

The small tag has a chance of being closed twice but is opened only once. This causes an error when trying to test HTML on a page that uses the created_updated template. It also causes the page that has both created and updated to show "Created on ..." to be small, but "Updated on..." to be normal size on the same line.

* Fix utilization graph on PowerFeeds within the Rack view.

Fixes #2099

* Rack Power Utilization proper calculation

Currently Rack Power Utilization does not take into account a device with utilization directly connected to a powerfeed. This adds the direct connected allocated_draw to the total.

* Add tests related to the Rack Power Utilization changes.