avoid crash when syncing a git repo without a worker (#4779)
If a worker is not available and you attempt to sync a git repo, a crash
will happen because you fall through to the case to redirecting based on
the job_result variable. But there is no job_result variable in scope at
this point. Instead go back to the original view.

Co-authored-by: Glenn Matthews <glenn.matthews@networktocode.com>