fix: passing pk for JobButton job_queue variable (#6791)
* fix: passing pk for JobButton job_queue variable

Original implemenation was passing the string name of the job_queue to be used. This was creating a payload that would not run the job on click as what is required is the UUID (pk) of the JobQueue object instead.

* fix: importing JobQueue model for changes

* fix: modifying unit test to check for pk

unit test originally was checking for the name value to exist in the hidden var for _job_queue against the rendering for job buttons. I have made the modifications to check for pk instead.

* Create 6770.fixed

high level details of change

---------

Co-authored-by: Hanlin Miao <46973263+HanlinMiao@users.noreply.github.com>
Co-authored-by: Glenn Matthews <glenn.matthews@networktocode.com>